requirements.txt
# Core ML libraries
numpy
pandas
scikit-learn
scipy

# Recommendation algorithms
scikit-surprise==1.1.3

# API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Experiment tracking
mlflow==2.9.2

# Caching
redis==5.0.1
hiredis==2.2.3

# Data processing
pyarrow==14.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Load testing
locust==2.18.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
tqdm==4.66.1

# Monitoring (optional, add later)
prometheus-client==0.19.0

# Development
black==23.12.1
flake8==6.1.0
ipython
jupyter==1.0.0
matplotlib
seaborn